+++ packages/react-mui-shared/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.58] - 2024-04-22
+
+### Added
+
+- Support optional onSearch callback for PangeaDataGrid. Triggered upon explicit refresh/search. Clicking the "Search" button or typing "Enter"
+
 ## [0.0.57] - 2024-03-25
 
 ### Fixed